improve useFieldArray tests structure
