+++ src/__tests__/useFieldArray.test.tsx
@@ -11,21 +11,13 @@ import { useFieldArray } from '../useFieldArray';
 import { useForm } from '../useForm';
 import * as generateId from '../logic/generateId';
 import { Controller } from '../controller';
-import { VALIDATION_MODE } from '../constants';
 import { FormProvider } from '../useFormContext';
-import {
-  Control,
-  FieldError,
-  DeepMap,
-  SubmitHandler,
-  UseFormMethods,
-  FieldValues,
-} from '../types';
+import { Control, SubmitHandler, UseFormMethods, FieldValues } from '../types';
 import isFunction from '../utils/isFunction';
 
 let nodeEnv: string | undefined;
 
-const mockGenerateId = () => {
+export const mockGenerateId = () => {
   let id = 0;
   jest.spyOn(generateId, 'default').mockImplementation(() => (id++).toString());
 };
@@ -859,3004 +851,6 @@ describe('useFieldArray', () => {
       expect(
         (screen.getByLabelText('test.0.firstName') as HTMLInputElement).value,
       ).toEqual('Bill');
-    });
-  });
-
-  describe('append', () => {
-    it('should append dirty fields correctly', async () => {
-      let dirtyInputs = {};
-      const Component = () => {
-        const {
-          register,
-          control,
-          formState: { dirty },
-        } = useForm<{
-          test: { value: string }[];
-        }>({
-          defaultValues: {
-            test: [
-              { value: 'plz change' },
-              { value: 'dont change' },
-              { value: 'dont change' },
-            ],
-          },
-        });
-        const { fields, append } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        dirtyInputs = dirty;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-            {dirty.test?.length && 'dirty'}
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.input(screen.getAllByRole('textbox')[0], {
-        target: { value: 'test' },
-      });
-      fireEvent.blur(screen.getAllByRole('textbox')[0]);
-
-      await waitFor(() => screen.getByText('dirty'));
-
-      expect(dirtyInputs).toEqual({
-        test: [{ value: true }],
-      });
-
-      fireEvent.click(screen.getByRole('button'));
-
-      expect(dirtyInputs).toEqual({
-        test: [{ value: true }, undefined, undefined, { value: true }],
-      });
-    });
-
-    it('should append data into the fields', () => {
-      let currentFields: any = [];
-      const Component = () => {
-        const { register, control } = useForm<{
-          test: { test: string }[];
-        }>();
-        const { fields, append } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        currentFields = fields;
-
-        return (
-          <form>
-            {fields.map((field, index) => {
-              return (
-                <input
-                  key={field.id}
-                  {...register(`test.${index}.test`)}
-                  defaultValue={field.test}
-                />
-              );
-            })}
-            <button type={'button'} onClick={() => append({ test: 'test' })}>
-              append
-            </button>
-            <button
-              type={'button'}
-              onClick={() =>
-                append([{ test: 'test-batch' }, { test: 'test-batch1' }])
-              }
-            >
-              appendBatch
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      act(() => {
-        fireEvent.click(screen.getByRole('button', { name: 'append' }));
-      });
-
-      act(() => {
-        expect(currentFields).toEqual([{ id: '0', test: 'test' }]);
-      });
-
-      act(() => {
-        fireEvent.click(screen.getByRole('button', { name: 'append' }));
-      });
-
-      act(() => {
-        expect(currentFields).toEqual([
-          { id: '0', test: 'test' },
-          { id: '1', test: 'test' },
-        ]);
-      });
-
-      act(() => {
-        fireEvent.click(screen.getByRole('button', { name: 'appendBatch' }));
-      });
-
-      act(() => {
-        expect(currentFields).toEqual([
-          { id: '0', test: 'test' },
-          { id: '1', test: 'test' },
-          { id: '2', test: 'test-batch' },
-          { id: '3', test: 'test-batch1' },
-        ]);
-      });
-    });
-
-    it.each(['isDirty', 'dirtyFields'])(
-      'should be dirty when value is appended with %s',
-      () => {
-        let isDirtyValue;
-        let dirtyValue;
-
-        const Component = () => {
-          const {
-            register,
-            control,
-            formState: { isDirty, dirty },
-          } = useForm<{
-            test: { test: string }[];
-          }>();
-          const { fields, append } = useFieldArray({
-            control,
-            name: 'test',
-          });
-
-          isDirtyValue = isDirty;
-          dirtyValue = dirty;
-
-          return (
-            <form>
-              {fields.map((field, index) => {
-                return (
-                  <input
-                    key={field.id}
-                    {...register(`test.${index}.test`)}
-                    defaultValue={field.test}
-                  />
-                );
-              })}
-              <button type={'button'} onClick={() => append({ test: 'test' })}>
-                append
-              </button>
-            </form>
-          );
-        };
-
-        render(<Component />);
-
-        act(() => {
-          fireEvent.click(screen.getByRole('button', { name: 'append' }));
-        });
-
-        act(() => {
-          fireEvent.click(screen.getByRole('button', { name: 'append' }));
-        });
-
-        act(() => {
-          fireEvent.click(screen.getByRole('button', { name: 'append' }));
-        });
-
-        expect(isDirtyValue).toBeTruthy();
-        expect(dirtyValue).toEqual({
-          test: [{ test: true }, { test: true }, { test: true }],
-        });
-      },
-    );
-
-    it('should trigger reRender when user is watching the all field array', () => {
-      const watched: any[] = [];
-      const Component = () => {
-        const { register, watch, control } = useForm<{
-          test: { value: string }[];
-        }>();
-        const { fields, append } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        watched.push(watch());
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      expect(watched).toEqual([
-        {},
-        {},
-        { test: [{ value: '' }] },
-        { test: [{ value: '' }] },
-      ]);
-    });
-
-    it('should focus if shouldFocus is true', () => {
-      const Component = () => {
-        const { register, control } = useForm<{
-          test: { value: string }[];
-        }>({
-          defaultValues: { test: [{ value: '1' }, { value: '2' }] },
-        });
-        const { fields, append } = useFieldArray({ control, name: 'test' });
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '3' })}>
-              append
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      const inputs = screen.getAllByRole('textbox');
-
-      expect(inputs).toHaveLength(3);
-
-      // expect(document.activeElement).toEqual(inputs[2]); no longer working
-    });
-
-    it('should not focus if shouldFocus is false', () => {
-      const Component = () => {
-        const { register, control } = useForm<{
-          test: { value: string }[];
-        }>({
-          defaultValues: { test: [{ value: '1' }, { value: '2' }] },
-        });
-        const { fields, append } = useFieldArray({ control, name: 'test' });
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-            <button
-              type="button"
-              onClick={() => append({ value: '3' }, { shouldFocus: false })}
-            >
-              append
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      const inputs = screen.getAllByRole('textbox');
-
-      expect(inputs).toHaveLength(3);
-      expect(document.activeElement).toEqual(document.body);
-    });
-
-    it('should return watched value with watch API', async () => {
-      const renderedItems: any = [];
-      const Component = () => {
-        const { watch, register, control } = useForm<{
-          test: { value: string }[];
-        }>();
-        const { fields, append } = useFieldArray({
-          name: 'test',
-          control,
-        });
-        const watched = watch('test');
-        renderedItems.push(watched);
-        return (
-          <div>
-            {fields.map((field, i) => (
-              <div key={field.id}>
-                <input
-                  defaultValue={field.value}
-                  {...register(`test.${i}.value` as const)}
-                />
-              </div>
-            ))}
-            <button onClick={() => append({ value: 'test' })}>append</button>
-          </div>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      await waitFor(() =>
-        expect(renderedItems).toEqual([
-          undefined,
-          undefined,
-          [{ value: 'test' }],
-          [{ value: 'test' }],
-        ]),
-      );
-    });
-
-    describe('with resolver', () => {
-      it('should invoke resolver when formState.isValid true', async () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-          });
-          const { append } = useFieldArray({ control, name: 'test' });
-          return { formState, append };
-        });
-
-        result.current.formState.isValid;
-
-        await act(async () => {
-          result.current.append({ value: '1' });
-        });
-
-        expect(resolver).toBeCalledWith(
-          {
-            test: [{ value: '1' }],
-          },
-          undefined,
-          false,
-        );
-      });
-
-      it('should not invoke resolver when formState.isValid false', () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-          });
-          const { append } = useFieldArray({ control, name: 'test' });
-          return { formState, append };
-        });
-
-        act(() => {
-          result.current.append({ value: '1' });
-        });
-
-        expect(resolver).not.toBeCalled();
-      });
-    });
-  });
-
-  describe('prepend', () => {
-    it('should pre-append data into the fields', () => {
-      let currentFields: any = [];
-
-      const Component = () => {
-        const { control, register } = useForm<{
-          test: {
-            test: string;
-          }[];
-        }>();
-        const { fields, prepend } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        currentFields = fields;
-
-        return (
-          <form>
-            {fields.map((field, index) => (
-              <div key={field.id}>
-                <input
-                  {...register(`test.${index}.test` as const)}
-                  defaultValue={field.test}
-                />
-              </div>
-            ))}
-            <button type={'button'} onClick={() => prepend({ test: 'test' })}>
-              prepend
-            </button>
-            <button
-              type={'button'}
-              onClick={() =>
-                prepend([{ test: 'test-batch' }, { test: 'test-batch1' }])
-              }
-            >
-              prependBatch
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      act(() => {
-        fireEvent.click(screen.getByRole('button', { name: 'prepend' }));
-      });
-
-      act(() => {
-        expect(currentFields).toEqual([{ id: '0', test: 'test' }]);
-      });
-
-      act(() => {
-        fireEvent.click(screen.getByRole('button', { name: 'prepend' }));
-      });
-
-      act(() => {
-        expect(currentFields).toEqual([
-          { id: '1', test: 'test' },
-          { id: '0', test: 'test' },
-        ]);
-      });
-
-      act(() => {
-        fireEvent.click(screen.getByRole('button', { name: 'prependBatch' }));
-      });
-
-      act(() => {
-        expect(currentFields).toEqual([
-          { id: '2', test: 'test-batch' },
-          { id: '3', test: 'test-batch1' },
-          { id: '1', test: 'test' },
-          { id: '0', test: 'test' },
-        ]);
-      });
-    });
-
-    it.each(['isDirty', 'dirtyFields'])(
-      'should be dirty when value is prepended with %s',
-      () => {
-        const { result } = renderHook(() => {
-          const { register, formState, control } = useForm();
-          const { fields, prepend } = useFieldArray({
-            control,
-            name: 'test',
-          });
-
-          return { register, formState, fields, prepend };
-        });
-
-        result.current.formState.isDirty;
-        result.current.formState.dirty;
-
-        act(() => {
-          result.current.prepend({ value: 'test' });
-        });
-
-        act(() => {
-          result.current.prepend({ value: 'test1' });
-        });
-
-        act(() => {
-          result.current.prepend({ value: 'test2' });
-        });
-
-        expect(result.current.formState.isDirty).toBeTruthy();
-        expect(result.current.formState.dirty).toEqual({
-          test: [{ value: true }, { value: true }, { value: true }],
-        });
-      },
-    );
-
-    it('should set prepended values to formState.touched', () => {
-      let touched: any;
-
-      const Component = () => {
-        const { register, formState, control } = useForm();
-        const { fields, prepend } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        touched = formState.touched;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input key={field.id} {...register(`test.${i}.value`)} />
-            ))}
-            <button
-              type="button"
-              onClick={() => prepend({ value: `test${1}` })}
-            >
-              prepend
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
-
-      fireEvent.blur(screen.getAllByRole('textbox')[0]);
-
-      fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
-      fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
-
-      expect(touched).toEqual({
-        test: [undefined, { value: true }, { value: true }],
-      });
-    });
-
-    it('should prepend error', async () => {
-      let errors: any;
-      const Component = () => {
-        const {
-          register,
-          formState: { errors: tempErrors },
-          handleSubmit,
-          control,
-        } = useForm<{
-          test: { value: string }[];
-        }>();
-        const { fields, prepend } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        errors = tempErrors;
-
-        return (
-          <form onSubmit={handleSubmit(() => {})}>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const, { required: true })}
-                defaultValue={field.value}
-              />
-            ))}
-            <button type="button" onClick={() => prepend({ value: '' })}>
-              prepend
-            </button>
-            <button>submit</button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
-
-      expect(errors.test).toBeUndefined();
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /submit/i }));
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
-
-      expect(errors.test).toHaveLength(2);
-    });
-
-    it('should trigger reRender when user is watching the all field array', () => {
-      const watched: any[] = [];
-      const Component = () => {
-        const { register, watch, control } = useForm<{
-          test: { value: string }[];
-        }>();
-        const { fields, prepend } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        watched.push(watch());
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-            <button type="button" onClick={() => prepend({ value: '' })}>
-              prepend
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: 'prepend' }));
-
-      expect(watched).toEqual([
-        {}, // first render
-        {}, // render inside useEffect in useFieldArray
-        { test: [{ value: '' }] }, // render inside prepend method
-        { test: [{ value: '' }] }, // render inside useEffect in useFieldArray
-      ]);
-    });
-
-    it('should return watched value with watch API', async () => {
-      const renderedItems: any = [];
-      const Component = () => {
-        const { watch, register, control } = useForm<{
-          test: {
-            value: string;
-          }[];
-        }>();
-        const { fields, append, prepend } = useFieldArray({
-          name: 'test',
-          control,
-        });
-        const watched = watch('test');
-        const isPrepended = React.useRef(false);
-        if (isPrepended.current) {
-          renderedItems.push(watched);
-        }
-        return (
-          <div>
-            {fields.map((field, i) => (
-              <div key={`${field.id}`}>
-                <input
-                  defaultValue={field.value}
-                  {...register(`test.${i}.value` as const)}
-                />
-              </div>
-            ))}
-            <button onClick={() => append({ value: '' })}>append</button>
-            <button
-              onClick={() => {
-                prepend({ value: 'test' });
-                isPrepended.current = true;
-              }}
-            >
-              prepend
-            </button>
-          </div>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      const inputs = screen.getAllByRole('textbox');
-
-      fireEvent.input(inputs[0], {
-        target: { name: 'test[0].value', value: '111' },
-      });
-      fireEvent.input(inputs[1], {
-        target: { name: 'test[1].value', value: '222' },
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
-
-      await waitFor(() =>
-        expect(renderedItems).toEqual([
-          [{ value: 'test' }, { value: '111' }, { value: '222' }],
-          [{ value: 'test' }, { value: '111' }, { value: '222' }],
-        ]),
-      );
-    });
-
-    it('should focus if shouldFocus is true', () => {
-      const Component = () => {
-        const { register, control } = useForm<{
-          test: { value: string }[];
-        }>({
-          defaultValues: {
-            test: [{ value: '1' }, { value: '2' }],
-          },
-        });
-        const { fields, prepend } = useFieldArray({ name: 'test', control });
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-            <button type="button" onClick={() => prepend({ value: '' })}>
-              prepend
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
-
-      const inputs = screen.getAllByRole('textbox');
-
-      expect(inputs).toHaveLength(3);
-      expect(document.activeElement).toEqual(inputs[0]);
-    });
-
-    it('should not focus if shouldFocus is false', () => {
-      const Component = () => {
-        const { register, control } = useForm({
-          defaultValues: {
-            test: [{ value: '1' }, { value: '2' }],
-          },
-        });
-        const { fields, prepend } = useFieldArray({ name: 'test', control });
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-            <button
-              type="button"
-              onClick={() => prepend({ value: '' }, { shouldFocus: false })}
-            >
-              prepend
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
-      const inputs = screen.getAllByRole('textbox');
-
-      expect(inputs).toHaveLength(3);
-      expect(document.activeElement).toEqual(document.body);
-    });
-
-    describe('with resolver', () => {
-      it('should invoke resolver when formState.isValid true', async () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-          });
-          const { prepend } = useFieldArray({ control, name: 'test' });
-          return { formState, prepend };
-        });
-
-        result.current.formState.isValid;
-
-        await act(async () => {
-          result.current.prepend({ value: '1' });
-        });
-
-        expect(resolver).toBeCalledWith(
-          {
-            test: [{ value: '1' }],
-          },
-          undefined,
-          false,
-        );
-      });
-
-      it('should not invoke resolver when formState.isValid false', () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-          });
-          const { prepend } = useFieldArray({ control, name: 'test' });
-          return { formState, prepend };
-        });
-
-        act(() => {
-          result.current.prepend({ value: '1' });
-        });
-
-        expect(resolver).not.toBeCalled();
-      });
-    });
-  });
-
-  describe('remove', () => {
-    it('should update isDirty formState when item removed', () => {
-      let formState: any;
-      const Component = () => {
-        const { register, control, formState: tempFormState } = useForm({
-          defaultValues: {
-            test: [{ name: 'default' }],
-          },
-        });
-        const { fields, remove, append } = useFieldArray({
-          name: 'test',
-          control,
-        });
-
-        formState = tempFormState;
-        formState.isDirty;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <div key={i.toString()}>
-                <input
-                  {...register(`test.${i}.name` as const)}
-                  defaultValue={field.name}
-                />
-                <button type={'button'} onClick={() => remove(i)}>
-                  remove
-                </button>
-              </div>
-            ))}
-
-            <button
-              type={'button'}
-              onClick={() =>
-                append({
-                  name: '',
-                })
-              }
-            >
-              append
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      expect(formState.isDirty).toBeFalsy();
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      expect(formState.isDirty).toBeTruthy();
-
-      fireEvent.click(screen.getAllByRole('button', { name: /remove/i })[1]);
-
-      expect(formState.isDirty).toBeFalsy();
-    });
-
-    it('should update isValid formState when item removed', async () => {
-      let formState: any;
-      const Component = () => {
-        const { register, control, formState: tempFormState } = useForm({
-          mode: 'onChange',
-          defaultValues: {
-            test: [{ name: 'default' }],
-          },
-        });
-        const { fields, remove, append } = useFieldArray({
-          name: 'test',
-          control,
-        });
-
-        formState = tempFormState;
-
-        formState.isValid;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <div key={i.toString()}>
-                <input
-                  {...register(`test.${i}.name` as const, { required: true })}
-                  defaultValue={field.name}
-                />
-                <button type={'button'} onClick={() => remove(i)}>
-                  remove
-                </button>
-              </div>
-            ))}
-
-            <button
-              type={'button'}
-              onClick={() =>
-                append({
-                  name: '',
-                })
-              }
-            >
-              append
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      await actComponent(async () => {
-        await fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      });
-
-      expect(formState.isValid).toBeFalsy();
-
-      await actComponent(async () => {
-        await fireEvent.click(
-          screen.getAllByRole('button', { name: /remove/i })[1],
-        );
-      });
-
-      expect(formState.isValid).toBeTruthy();
-    });
-
-    it('should remove field according index', () => {
-      const { result } = renderHook(() => {
-        const { control } = useForm({
-          defaultValues: {
-            test: [{ value: 'default' }],
-          },
-        });
-        return useFieldArray({
-          control,
-          name: 'test',
-        });
-      });
-
-      act(() => {
-        result.current.append({ value: 'test' });
-      });
-
-      act(() => {
-        result.current.remove(1);
-      });
-
-      expect(result.current.fields).toEqual([{ id: '0', value: 'default' }]);
-
-      act(() => {
-        result.current.remove(0);
-      });
-
-      expect(result.current.fields).toEqual([]);
-    });
-
-    it('should remove all field', () => {
-      const { result } = renderHook(() => {
-        const { control } = useForm({
-          defaultValues: {
-            test: [{ value: 'default' }],
-          },
-        });
-        return useFieldArray({
-          control,
-          name: 'test',
-        });
-      });
-
-      act(() => {
-        result.current.append({ value: 'test' });
-      });
-
-      act(() => {
-        result.current.remove();
-      });
-
-      expect(result.current.fields).toEqual([]);
-    });
-
-    it('should remove specific fields when index is array', () => {
-      const { result } = renderHook(() => {
-        const { control } = useForm({
-          defaultValues: {
-            test: [{ value: 'default' }],
-          },
-        });
-        return useFieldArray({
-          control,
-          name: 'test',
-        });
-      });
-
-      act(() => {
-        result.current.append({ value: 'test' });
-      });
-
-      act(() => {
-        result.current.remove([0, 1]);
-      });
-
-      expect(result.current.fields).toEqual([]);
-    });
-
-    it.each(['isDirty', 'dirtyFields'])(
-      'should be dirty when value is remove with %s',
-      () => {
-        const { result } = renderHook(() => {
-          const { register, formState, control } = useForm({
-            defaultValues: {
-              test: [{ value: 'default' }],
-            },
-          });
-          const { fields, append, remove } = useFieldArray({
-            control,
-            name: 'test',
-          });
-
-          return { register, formState, fields, append, remove };
-        });
-
-        result.current.formState.isDirty;
-        result.current.formState.dirty;
-
-        act(() => {
-          result.current.append({ value: 'test' });
-        });
-
-        act(() => {
-          result.current.append({ value: 'test' });
-        });
-
-        act(() => {
-          result.current.remove(0);
-        });
-
-        expect(result.current.formState.isDirty).toBeTruthy();
-        expect(result.current.formState.dirty).toEqual({
-          test: [{ value: true }, { value: true }],
-        });
-
-        act(() => {
-          result.current.remove();
-        });
-
-        expect(result.current.formState.isDirty).toBeTruthy();
-        expect(result.current.formState.dirty).toEqual({
-          test: [{ value: true }],
-        });
-      },
-    );
-
-    it('should remove values from formState.touched', () => {
-      let touched: any;
-
-      const Component = () => {
-        const { register, formState, control } = useForm();
-        const { fields, append, remove } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        touched = formState.touched;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input key={field.id} {...register(`test.${i}.value`)} />
-            ))}
-            <button type="button" onClick={() => append({ value: 'append' })}>
-              append
-            </button>
-            <button type="button" onClick={() => remove(0)}>
-              remove
-            </button>
-            <button type="button" onClick={() => remove()}>
-              remove all
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      const inputs = screen.getAllByRole('textbox');
-
-      fireEvent.blur(inputs[0]);
-      fireEvent.blur(inputs[1]);
-      fireEvent.blur(inputs[2]);
-
-      expect(touched).toEqual({
-        test: [{ value: true }, { value: true }, { value: true }],
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: 'remove' }));
-
-      expect(touched).toEqual({
-        test: [{ value: true }, { value: true }],
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: 'remove all' }));
-
-      expect(touched).toEqual({});
-    });
-
-    it('should remove specific field if isValid is true', async () => {
-      let isValid: any;
-      const Component = () => {
-        const { register, formState, control } = useForm({
-          mode: VALIDATION_MODE.onChange,
-        });
-        const { fields, append, remove } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        isValid = formState.isValid;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const, { required: true })}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-            <button type="button" onClick={() => remove(1)}>
-              remove
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      });
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      });
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      });
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      });
-
-      expect(isValid).toBeFalsy();
-
-      const inputs = screen.getAllByRole('textbox');
-
-      await actComponent(async () => {
-        fireEvent.input(inputs[0], {
-          target: { value: 'test' },
-        });
-      });
-
-      await actComponent(async () => {
-        fireEvent.input(inputs[2], {
-          target: { value: 'test' },
-        });
-      });
-
-      await actComponent(async () => {
-        fireEvent.input(inputs[3], {
-          target: { value: 'test' },
-        });
-      });
-
-      expect(isValid).toBeFalsy();
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: 'remove' }));
-      });
-
-      expect(isValid).toBeTruthy();
-    });
-
-    it('should remove all field if isValid is true', async () => {
-      let isValid: any;
-      const Component = () => {
-        const { register, formState, control } = useForm({
-          mode: VALIDATION_MODE.onChange,
-        });
-        const { fields, append, remove } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        isValid = formState.isValid;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const, { required: true })}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-            <button type="button" onClick={() => remove()}>
-              remove
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      });
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      });
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      });
-
-      expect(isValid).toBeFalsy();
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: 'remove' }));
-      });
-
-      expect(isValid).toBeTruthy();
-    });
-
-    it('should remove error', async () => {
-      let errors: any;
-      const Component = () => {
-        const {
-          register,
-          formState: { errors: tempErrors },
-          handleSubmit,
-          control,
-        } = useForm();
-        const { fields, append, remove } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        errors = tempErrors;
-
-        return (
-          <form onSubmit={handleSubmit(() => {})}>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const, { required: true })}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-            <button type="button" onClick={() => remove(0)}>
-              remove
-            </button>
-            <button type="button" onClick={() => remove()}>
-              remove all
-            </button>
-            <button>submit</button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /submit/i }));
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: 'remove' }));
-
-      expect(errors.test).toHaveLength(2);
-
-      fireEvent.click(screen.getByRole('button', { name: 'remove all' }));
-
-      expect(errors.test).toBeUndefined();
-    });
-
-    it('should remove nested field array error', async () => {
-      type FormValues = {
-        test: {
-          nested: {
-            test: string;
-            key: number;
-          }[];
-        }[];
-      };
-
-      let mockKey = 0;
-      const Nested = ({
-        register,
-        errors,
-        control,
-        index,
-      }: {
-        register: UseFormMethods['register'];
-        control: Control<FormValues>;
-        errors: DeepMap<Record<string, any>, FieldError>;
-        index: number;
-      }) => {
-        const { fields, append, remove } = useFieldArray({
-          name: `test.${index}.nested` as any,
-          control,
-        });
-        return (
-          <fieldset>
-            {fields.map((field, i) => (
-              <div key={field.id}>
-                <input
-                  {...register(`test.${index}.nested.${i}.test`, {
-                    required: 'required',
-                  })}
-                />
-                {errors?.test &&
-                  errors.test[index]?.nested &&
-                  errors.test[index].nested[i]?.test && (
-                    <span data-testid="nested-error">
-                      {errors.test[index].nested[i].test.message}
-                    </span>
-                  )}
-                <button onClick={() => remove(i)}>nested delete</button>
-              </div>
-            ))}
-            <button onClick={() => append({ test: 'test', key: mockKey++ })}>
-              nested append
-            </button>
-          </fieldset>
-        );
-      };
-      const callback = jest.fn();
-      const Component = () => {
-        const {
-          register,
-          formState: { errors },
-          handleSubmit,
-          control,
-        } = useForm<FormValues>({
-          defaultValues: {
-            test: [{ nested: [{ test: '', key: mockKey }] }],
-          },
-        });
-        const { fields } = useFieldArray({ name: 'test', control });
-        return (
-          <form onSubmit={handleSubmit(callback)}>
-            {fields.map((_, i) => (
-              <Nested
-                key={i.toString()}
-                // @ts-ignore
-                register={register}
-                errors={errors}
-                control={control}
-                index={i}
-              />
-            ))}
-            <button>submit</button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      await actComponent(async () => {
-        await fireEvent.click(screen.getByRole('button', { name: /submit/i }));
-      });
-
-      expect(screen.queryByTestId('nested-error')).toBeInTheDocument();
-
-      await actComponent(async () => {
-        await fireEvent.click(
-          screen.getByRole('button', { name: /nested delete/i }),
-        );
-      });
-
-      expect(screen.queryByTestId('nested-error')).not.toBeInTheDocument();
-
-      await actComponent(async () => {
-        await fireEvent.click(
-          screen.getByRole('button', { name: /nested append/i }),
-        );
-      });
-
-      expect(screen.queryByTestId('nested-error')).not.toBeInTheDocument();
-    });
-
-    it('should trigger reRender when user is watching the all field array', () => {
-      const watched: any[] = [];
-      const Component = () => {
-        const { register, watch, control } = useForm<{
-          test: {
-            value: string;
-          }[];
-        }>();
-        const { fields, append, remove } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        watched.push(watch());
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                defaultValue={field.value}
-                {...register(`test.${i}.value` as const)}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-            <button type="button" onClick={() => remove(0)}>
-              remove
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: 'append' }));
-
-      fireEvent.click(screen.getByRole('button', { name: 'remove' }));
-
-      expect(watched).toEqual([
-        {}, // first render
-        {}, // render inside useEffect in useFieldArray
-        { test: [{ value: '' }] }, // render inside append method
-        { test: [{ value: '' }] }, // render inside useEffect in useFieldArray
-        { test: [] }, // render inside remove method
-        {}, // render inside useEffect in useFieldArray
-      ]);
-    });
-
-    it('should return watched value with watch API', async () => {
-      const renderedItems: any = [];
-      const Component = () => {
-        const { watch, register, control } = useForm<{
-          test: {
-            value: string;
-          }[];
-        }>();
-        const { fields, append, remove } = useFieldArray({
-          name: 'test',
-          control,
-        });
-        const watched = watch('test');
-        const isRemoved = React.useRef(false);
-        if (isRemoved.current) {
-          renderedItems.push(watched);
-        }
-
-        return (
-          <div>
-            {fields.map((field, i) => (
-              <div key={`${field.id}`}>
-                <input
-                  {...register(`test.${i}.value` as const)}
-                  defaultValue={field.value}
-                />
-              </div>
-            ))}
-            <button onClick={() => append({ value: '' })}>append</button>
-            <button
-              onClick={() => {
-                remove(2);
-                isRemoved.current = true;
-              }}
-            >
-              remove
-            </button>
-          </div>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      const inputs = screen.getAllByRole('textbox');
-
-      fireEvent.change(inputs[0], {
-        target: { name: 'test[0].value', value: '111' },
-      });
-      fireEvent.change(inputs[1], {
-        target: { name: 'test[1].value', value: '222' },
-      });
-      fireEvent.change(inputs[2], {
-        target: { name: 'test[2].value', value: '333' },
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: /remove/i }));
-
-      await waitFor(() =>
-        expect(renderedItems).toEqual([
-          [{ value: '111' }, { value: '222' }],
-          [{ value: '111' }, { value: '222' }],
-        ]),
-      );
-    });
-
-    it('should remove dirty fields with nested field inputs', () => {
-      const { result } = renderHook(() => {
-        const { register, formState, control } = useForm({
-          defaultValues: {
-            test: {
-              data: [{ value: 'default' }],
-            },
-          },
-        });
-        const { fields, append, remove } = useFieldArray({
-          control,
-          name: 'test.data',
-        });
-
-        return { register, formState, fields, append, remove };
-      });
-
-      result.current.formState.dirty as Record<string, any>;
-      result.current.formState.isDirty;
-
-      act(() => {
-        result.current.append({ value: 'test' });
-      });
-
-      expect(result.current.formState.isDirty).toBeTruthy();
-      expect(result.current.formState.dirty).toEqual({
-        test: { data: [undefined, { value: true }] },
-      });
-
-      act(() => {
-        result.current.remove(1);
-      });
-
-      expect(result.current.formState.isDirty).toBeFalsy();
-      expect(result.current.formState.dirty).toEqual({});
-    });
-
-    it('should remove Controller by index without error', () => {
-      const Component = () => {
-        const { control, handleSubmit } = useForm<{
-          test: {
-            firstName: string;
-          }[];
-        }>({
-          defaultValues: {
-            test: [],
-          },
-        });
-        const { fields, append, remove } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        return (
-          <form onSubmit={handleSubmit(() => {})}>
-            <ul>
-              {fields.map((item, index) => {
-                return (
-                  <li key={item.id}>
-                    <Controller
-                      render={({ field }) => <input {...field} />}
-                      name={`test.${index}.firstName` as const}
-                      control={control}
-                      defaultValue={item.firstName}
-                    />
-                    <button type="button" onClick={() => remove(index)}>
-                      delete
-                    </button>
-                  </li>
-                );
-              })}
-            </ul>
-            <section>
-              <button
-                type="button"
-                onClick={() => {
-                  append({ firstName: 'appendBill' });
-                }}
-              >
-                append
-              </button>
-            </section>
-
-            <input type="submit" />
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      actComponent(() => {
-        fireEvent.click(screen.getByRole('button', { name: 'append' }));
-      });
-      actComponent(() => {
-        fireEvent.click(screen.getByRole('button', { name: 'append' }));
-      });
-      actComponent(() => {
-        fireEvent.click(screen.getByRole('button', { name: 'append' }));
-      });
-      actComponent(() => {
-        fireEvent.click(screen.getByRole('button', { name: 'append' }));
-      });
-
-      actComponent(() => {
-        fireEvent.click(screen.getAllByRole('button', { name: 'delete' })[1]);
-      });
-      actComponent(() => {
-        fireEvent.click(screen.getAllByRole('button', { name: 'delete' })[1]);
-      });
-      actComponent(() => {
-        fireEvent.click(screen.getAllByRole('button', { name: 'delete' })[1]);
-      });
-      actComponent(() => {
-        fireEvent.click(screen.getAllByRole('button', { name: 'delete' })[0]);
-      });
-    });
-
-    it("should not reset Controller's value during remove when Field Array name is already registered", () => {
-      function Component() {
-        const { control, handleSubmit } = useForm({
-          defaultValues: {
-            test: [{ firstName: 'Bill', lastName: '' }],
-          },
-        });
-        const { fields, append, remove } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        return (
-          <form onSubmit={handleSubmit(() => {})}>
-            <ul>
-              {fields.map((item, index) => {
-                return (
-                  <li key={item.id}>
-                    <Controller
-                      name={`test.${index}.lastName` as const}
-                      control={control}
-                      defaultValue={item.lastName} // make sure to set up defaultValue
-                      render={(props: any) => <input {...props} />}
-                    />
-                    <button type="button" onClick={() => remove(index)}>
-                      Delete
-                    </button>
-                  </li>
-                );
-              })}
-            </ul>
-            <button
-              type="button"
-              onClick={() => {
-                append({ firstName: 'appendBill', lastName: 'appendLuo' });
-              }}
-            >
-              append
-            </button>
-          </form>
-        );
-      }
-
-      render(<Component />);
-
-      fireEvent.input(screen.getAllByRole('textbox')[0], {
-        target: { name: 'test[0].lastName', value: '111' },
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: 'append' }));
-
-      fireEvent.click(screen.getAllByRole('button', { name: 'Delete' })[1]);
-
-      expect(
-        (screen.getAllByRole('textbox')[0] as HTMLInputElement).value,
-      ).toEqual('111');
-    });
-
-    describe('with resolver', () => {
-      it('should invoke resolver when formState.isValid true', async () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-            defaultValues: {
-              test: [{ value: 'test' }],
-            },
-          });
-          const { remove } = useFieldArray({ control, name: 'test' });
-          return { formState, remove };
-        });
-
-        result.current.formState.isValid;
-
-        await act(async () => {
-          result.current.remove(0);
-        });
-
-        expect(resolver).toBeCalledWith(
-          {
-            test: [],
-          },
-          undefined,
-          false,
-        );
-      });
-
-      it('should not invoke resolver when formState.isValid false', () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-            defaultValues: {
-              test: [{ value: 'test' }],
-            },
-          });
-          const { remove } = useFieldArray({ control, name: 'test' });
-          return { formState, remove };
-        });
-
-        act(() => {
-          result.current.remove(0);
-        });
-
-        expect(resolver).not.toBeCalled();
-      });
-    });
-  });
-
-  describe('insert', () => {
-    it('should insert data at index with single value', () => {
-      const { result } = renderHook(() => {
-        const { control } = useForm({
-          defaultValues: {
-            test: [{ test: '1' }, { test: '2' }],
-          },
-        });
-        const { fields, insert } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        return { fields, insert };
-      });
-
-      act(() => {
-        result.current.insert(1, { test: '3' });
-      });
-
-      expect(result.current.fields).toEqual([
-        { id: '0', test: '1' },
-        { id: '2', test: '3' },
-        { id: '1', test: '2' },
-      ]);
-    });
-
-    it('should insert data at index with array value', () => {
-      const { result } = renderHook(() => {
-        const { control } = useForm({
-          defaultValues: {
-            test: [{ test: '1' }, { test: '2' }],
-          },
-        });
-        const { fields, insert } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        return { fields, insert };
-      });
-
-      act(() => {
-        result.current.insert(1, [{ test: '3' }, { test: '4' }]);
-      });
-
-      expect(result.current.fields).toEqual([
-        { id: '0', test: '1' },
-        { id: '2', test: '3' },
-        { id: '3', test: '4' },
-        { id: '1', test: '2' },
-      ]);
-    });
-
-    it.each(['isDirty', 'dirtyFields'])(
-      'should insert data to formState.%s at index with single value',
-      () => {
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm<{
-            test: { value: string; value1: string }[];
-          }>({
-            defaultValues: { test: [{ value: '1' }] },
-          });
-          const { fields, append, insert } = useFieldArray({
-            control,
-            name: 'test',
-          });
-
-          return { formState, fields, append, insert };
-        });
-
-        result.current.formState.isDirty;
-        result.current.formState.dirty;
-
-        act(() => {
-          result.current.append({ value: '2' });
-        });
-
-        act(() => {
-          result.current.insert(1, { value1: '3' });
-        });
-
-        expect(result.current.formState.isDirty).toBeTruthy();
-        expect(result.current.formState.dirty).toEqual({
-          test: [undefined, { value: true, value1: true }, { value: true }],
-        });
-      },
-    );
-
-    it.each(['isDirty', 'dirtyFields'])(
-      'should insert data to formState.%s at index with array value',
-      () => {
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm<{
-            test: { value1: string; value2: string; value: string }[];
-          }>({
-            defaultValues: { test: [{ value: '1' }] },
-          });
-          const { fields, append, insert } = useFieldArray({
-            control,
-            name: 'test',
-          });
-
-          return { formState, fields, append, insert };
-        });
-
-        result.current.formState.isDirty;
-        result.current.formState.dirty;
-
-        act(() => {
-          result.current.append({ value: '2' });
-        });
-
-        act(() => {
-          result.current.insert(1, [{ value1: '3' }, { value2: '4' }]);
-        });
-
-        expect(result.current.formState.isDirty).toBeTruthy();
-        expect(result.current.formState.dirty).toEqual({
-          test: [
-            undefined,
-            { value1: true, value: true },
-            { value2: true, value: true },
-            { value: true },
-          ],
-        });
-      },
-    );
-
-    it('should insert touched fields with single value', () => {
-      let touched: any;
-      const Component = () => {
-        const { register, formState, control } = useForm({
-          defaultValues: { test: [{ value: '1' }, { value: '2' }] },
-        });
-        const { fields, insert } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        touched = formState.touched;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                defaultValue={field.value}
-                {...register(`test.${i}.value` as const)}
-              />
-            ))}
-            <button
-              type="button"
-              onClick={() => insert(1, { value: `${fields.length}` })}
-            >
-              insert
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.blur(screen.getAllByRole('textbox')[0]);
-      fireEvent.blur(screen.getAllByRole('textbox')[1]);
-
-      fireEvent.click(screen.getByRole('button', { name: /insert/i }));
-
-      expect(touched).toEqual({
-        test: [{ value: true }, undefined, { value: true }],
-      });
-    });
-
-    it('should insert touched fields with array value', () => {
-      let touched: any;
-      const Component = () => {
-        const { register, formState, control } = useForm({
-          defaultValues: { test: [{ value: '1' }, { value: '2' }] },
-        });
-        const { fields, insert } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        touched = formState.touched;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                defaultValue={field.value}
-                {...register(`test.${i}.value` as const)}
-              />
-            ))}
-            <button
-              type="button"
-              onClick={() =>
-                insert(1, [
-                  { value: `${fields.length}` },
-                  { value: `${fields.length + 1}` },
-                ])
-              }
-            >
-              insert array
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.blur(screen.getAllByRole('textbox')[0]);
-      fireEvent.blur(screen.getAllByRole('textbox')[1]);
-
-      fireEvent.click(screen.getByRole('button', { name: /insert array/i }));
-
-      expect(touched).toEqual({
-        test: [{ value: true }, undefined, undefined, { value: true }],
-      });
-    });
-
-    it('should insert error with single value', async () => {
-      let errors: any;
-      const Component = () => {
-        const { register, handleSubmit, control, ...rest } = useForm();
-        const { fields, append, insert } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        errors = rest.formState.errors;
-
-        return (
-          <form onSubmit={handleSubmit(() => {})}>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value`, { required: true })}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-            <button type="button" onClick={() => insert(1, { value: '' })}>
-              insert
-            </button>
-            <button>submit</button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /submit/i }));
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: /insert/i }));
-
-      expect(errors.test[0]).toBeDefined();
-      expect(errors.test[1]).toBeUndefined();
-      expect(errors.test[2]).toBeDefined();
-    });
-
-    it('should insert error with array value', async () => {
-      let errors: any;
-      const Component = () => {
-        const { register, handleSubmit, control, ...rest } = useForm();
-        const { fields, append, insert } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        errors = rest.formState.errors;
-
-        return (
-          <form onSubmit={handleSubmit(() => {})}>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value`, { required: true })}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-            <button
-              type="button"
-              onClick={() => insert(1, [{ value: '' }, { value: '' }])}
-            >
-              insert array
-            </button>
-            <button>submit</button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /submit/i }));
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: /insert array/i }));
-
-      expect(errors.test[0]).toBeDefined();
-      expect(errors.test[1]).toBeUndefined();
-      expect(errors.test[2]).toBeUndefined();
-      expect(errors.test[3]).toBeDefined();
-    });
-
-    it('should focus if shouldFocus is true', () => {
-      const Component = () => {
-        const { register, control } = useForm({
-          defaultValues: {
-            test: [{ value: '1' }, { value: '2' }],
-          },
-        });
-        const { fields, insert } = useFieldArray({ name: 'test', control });
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-            <button type="button" onClick={() => insert(1, { value: '' })}>
-              insert
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /insert/i }));
-
-      const inputs = screen.getAllByRole('textbox');
-
-      expect(inputs).toHaveLength(3);
-      expect(document.activeElement).toEqual(inputs[1]);
-    });
-
-    it('should not focus if shouldFocus is false', () => {
-      const Component = () => {
-        const { register, control } = useForm({
-          defaultValues: {
-            test: [{ value: '1' }, { value: '2' }],
-          },
-        });
-        const { fields, insert } = useFieldArray({ name: 'test', control });
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-            <button
-              type="button"
-              onClick={() => insert(1, { value: '' }, { shouldFocus: false })}
-            >
-              insert
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /insert/i }));
-
-      const inputs = screen.getAllByRole('textbox');
-
-      expect(inputs).toHaveLength(3);
-      expect(document.activeElement).toEqual(document.body);
-    });
-
-    it('should trigger reRender when user is watching the all field array', () => {
-      const watched: any[] = [];
-      const Component = () => {
-        const { register, watch, control } = useForm<{
-          test: {
-            value: string;
-          }[];
-        }>();
-        const { fields, insert } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        watched.push(watch());
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                defaultValue={field.value}
-                {...register(`test.${i}.value` as const)}
-              />
-            ))}
-            <button type="button" onClick={() => insert(0, { value: '' })}>
-              insert
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: 'insert' }));
-
-      expect(watched).toEqual([
-        {}, // first render
-        {}, // render inside useEffect in useFieldArray
-        { test: [{ value: '' }] }, // render inside insert method
-        { test: [{ value: '' }] }, // render inside useEffect in useFieldArray
-      ]);
-    });
-
-    it('should return watched value with watch API', async () => {
-      const renderedItems: any = [];
-      const Component = () => {
-        const { watch, register, control } = useForm<{
-          test: {
-            value: string;
-          }[];
-        }>();
-        const { fields, append, insert } = useFieldArray({
-          name: 'test',
-          control,
-        });
-        const watched = watch('test');
-        const isInserted = React.useRef(false);
-        if (isInserted.current) {
-          renderedItems.push(watched);
-        }
-        return (
-          <div>
-            {fields.map((field, i) => (
-              <div key={`${field.id}`}>
-                <input
-                  defaultValue={field.value}
-                  {...register(`test.${i}.value` as const)}
-                />
-              </div>
-            ))}
-            <button onClick={() => append({ value: '' })}>append</button>
-            <button
-              onClick={() => {
-                insert(1, { value: 'test' });
-                isInserted.current = true;
-              }}
-            >
-              insert
-            </button>
-          </div>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      const inputs = screen.getAllByRole('textbox');
-
-      fireEvent.change(inputs[0], {
-        target: { name: 'test[0].value', value: '111' },
-      });
-      fireEvent.change(inputs[1], {
-        target: { name: 'test[1].value', value: '222' },
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: /insert/i }));
-
-      expect(renderedItems).toEqual([
-        [{ value: '111' }, { value: 'test' }, { value: '222' }],
-        [{ value: '111' }, { value: 'test' }, { value: '222' }],
-      ]);
-    });
-
-    describe('with resolver', () => {
-      it('should invoke resolver when formState.isValid true', async () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-          });
-          const { insert } = useFieldArray({ control, name: 'test' });
-          return { formState, insert };
-        });
-
-        result.current.formState.isValid;
-
-        await act(async () => {
-          result.current.insert(0, { value: '1' });
-        });
-
-        expect(resolver).toBeCalledWith(
-          {
-            test: [{ value: '1' }],
-          },
-          undefined,
-          false,
-        );
-      });
-
-      it('should not invoke resolver when formState.isValid false', () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-          });
-          const { insert } = useFieldArray({ control, name: 'test' });
-          return { formState, insert };
-        });
-
-        act(() => {
-          result.current.insert(0, { value: '1' });
-        });
-
-        expect(resolver).not.toBeCalled();
-      });
-    });
-  });
-
-  describe('swap', () => {
-    it('should swap into pointed position', () => {
-      const { result } = renderHook(() => {
-        const { register, control } = useForm({
-          defaultValues: { test: [{ value: '1' }] },
-        });
-        const methods = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        return { register, ...methods };
-      });
-
-      act(() => {
-        result.current.append({ value: '2' });
-      });
-
-      act(() => {
-        result.current.swap(0, 1);
-      });
-
-      expect(result.current.fields).toEqual([
-        { id: '1', value: '2' },
-        { id: '0', value: '1' },
-      ]);
-    });
-
-    it('should swap data order', () => {
-      const { result } = renderHook(() => {
-        const { register, control } = useForm({
-          defaultValues: { test: [{ value: '1' }] },
-        });
-        const methods = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        return { register, ...methods };
-      });
-
-      act(() => {
-        result.current.append({ value: '2' });
-      });
-
-      act(() => {
-        result.current.swap(0, 1);
-      });
-
-      expect(result.current.fields).toEqual([
-        { id: '1', value: '2' },
-        { id: '0', value: '1' },
-      ]);
-    });
-
-    it.each(['isDirty', 'dirtyFields'])(
-      'should swap dirty order when formState.%s is defined',
-      () => {
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            defaultValues: { test: [{ value: '1' }] },
-          });
-          const methods = useFieldArray({
-            control,
-            name: 'test',
-          });
-          return {
-            formState,
-            ...methods,
-          };
-        });
-
-        result.current.formState.isDirty;
-        result.current.formState.dirty;
-
-        act(() => {
-          result.current.append({ value: '2' });
-        });
-
-        act(() => {
-          result.current.append({ value: '3' });
-        });
-
-        act(() => {
-          result.current.swap(0, 1);
-        });
-
-        expect(result.current.formState.isDirty).toBeTruthy();
-        expect(result.current.formState.dirty).toEqual({
-          test: [{ value: true }, { value: true }, { value: true }],
-        });
-      },
-    );
-
-    it('should swap errors', async () => {
-      let errors: any;
-      const Component = () => {
-        const { register, handleSubmit, control, ...rest } = useForm({
-          defaultValues: { test: [{ value: 'test' }] },
-        });
-        const { fields, append, swap } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        errors = rest.formState.errors;
-
-        return (
-          <form onSubmit={handleSubmit(() => {})}>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const, { required: true })}
-                defaultValue={field.value}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-            <button type="button" onClick={() => swap(0, 1)}>
-              swap
-            </button>
-            <button>submit</button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /submit/i }));
-      });
-
-      expect(errors.test[0]).toBeUndefined();
-      expect(errors.test[1]).toBeDefined();
-
-      fireEvent.click(screen.getByRole('button', { name: /swap/i }));
-
-      expect(errors.test[0]).toBeDefined();
-      expect(errors.test[1]).toBeUndefined();
-    });
-
-    it('should swap touched fields', async () => {
-      let touched: any;
-      const Component = () => {
-        const { register, formState, control } = useForm({
-          defaultValues: { test: [{ value: 'test' }] },
-        });
-        const { fields, append, swap } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        touched = formState.touched;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-            <button type="button" onClick={() => swap(0, 1)}>
-              swap
-            </button>
-            <button>submit</button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      fireEvent.blur(screen.getAllByRole('textbox')[0]);
-
-      fireEvent.click(screen.getByRole('button', { name: /swap/i }));
-
-      expect(touched).toEqual({
-        test: [undefined, { value: true }],
-      });
-    });
-
-    it('should trigger reRender when user is watching the all field array', () => {
-      const watched: any[] = [];
-      const Component = () => {
-        const { register, watch, control } = useForm({
-          defaultValues: {
-            test: [{ value: '1' }, { value: '2' }],
-          },
-        });
-        const { fields, swap } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        watched.push(watch());
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                defaultValue={field.value}
-                {...register(`test.${i}.value` as const)}
-              />
-            ))}
-            <button type="button" onClick={() => swap(0, 1)}>
-              swap
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: 'swap' }));
-
-      expect(watched).toEqual([
-        { test: [{ value: '1' }, { value: '2' }] }, // first render
-        { test: [{ value: '1' }, { value: '2' }] }, // render inside useEffect in useFieldArray
-        { test: [{ value: '2' }, { value: '1' }] }, // render inside swap method
-        { test: [{ value: '2' }, { value: '1' }] }, // render inside useEffect in useFieldArray
-      ]);
-    });
-
-    it('should return watched value with watch API', async () => {
-      const renderedItems: any = [];
-      const Component = () => {
-        const { watch, register, control } = useForm<{
-          test: {
-            value: string;
-          }[];
-        }>();
-        const { fields, append, swap } = useFieldArray({
-          name: 'test',
-          control,
-        });
-        const watched = watch('test');
-        const isSwapped = React.useRef(false);
-        if (isSwapped.current) {
-          renderedItems.push(watched);
-        }
-        return (
-          <div>
-            {fields.map((field, i) => (
-              <div key={`${field.id}`}>
-                <input
-                  defaultValue={field.value}
-                  {...register(`test.${i}.value` as const)}
-                />
-              </div>
-            ))}
-            <button onClick={() => append({ value: '' })}>append</button>
-            <button
-              onClick={() => {
-                swap(0, 1);
-                isSwapped.current = true;
-              }}
-            >
-              swap
-            </button>
-          </div>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      const inputs = screen.getAllByRole('textbox');
-
-      fireEvent.input(inputs[0], {
-        target: { name: 'test[0].value', value: '111' },
-      });
-      fireEvent.input(inputs[1], {
-        target: { name: 'test[1].value', value: '222' },
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: /swap/i }));
-
-      await waitFor(() =>
-        expect(renderedItems).toEqual([
-          [{ value: '222' }, { value: '111' }],
-          [{ value: '222' }, { value: '111' }],
-        ]),
-      );
-    });
-
-    describe('with resolver', () => {
-      it('should invoke resolver when formState.isValid true', async () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-            defaultValues: {
-              test: [{ value: '1' }, { value: '2' }],
-            },
-          });
-          const { swap } = useFieldArray({ control, name: 'test' });
-          return { formState, swap };
-        });
-
-        result.current.formState.isValid;
-
-        await act(async () => {
-          result.current.swap(0, 1);
-        });
-
-        expect(resolver).toBeCalledWith(
-          {
-            test: [{ value: '2' }, { value: '1' }],
-          },
-          undefined,
-          false,
-        );
-      });
-
-      it('should not invoke resolver when formState.isValid false', () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-            defaultValues: {
-              test: [{ value: '1' }, { value: '2' }],
-            },
-          });
-          const { swap } = useFieldArray({ control, name: 'test' });
-          return { formState, swap };
-        });
-
-        act(() => {
-          result.current.swap(0, 1);
-        });
-
-        expect(resolver).not.toBeCalled();
-      });
-    });
-  });
-
-  describe('move', () => {
-    it.each(['isDirty', 'dirty'])(
-      'should move dirty into pointed position when formState.%s is defined',
-      () => {
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            defaultValues: { test: [{ value: '1' }] },
-          });
-          const methods = useFieldArray({
-            control,
-            name: 'test',
-          });
-          return {
-            formState,
-            ...methods,
-          };
-        });
-
-        result.current.formState.isDirty;
-        result.current.formState.dirty;
-
-        act(() => {
-          result.current.append({ value: '2' });
-        });
-
-        act(() => {
-          result.current.append({ value: '3' });
-        });
-
-        act(() => {
-          result.current.move(0, 1);
-        });
-
-        expect(result.current.formState.isDirty).toBeTruthy();
-        expect(result.current.formState.dirty).toEqual({
-          test: [{ value: true }, { value: true }, { value: true }],
-        });
-      },
-    );
-
-    it('should move errors', async () => {
-      let errors: any;
-      const Component = () => {
-        const { register, handleSubmit, control, ...rest } = useForm({
-          defaultValues: { test: [{ value: 'test' }] },
-        });
-        const { fields, append, move } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        errors = rest.formState.errors;
-
-        return (
-          <form onSubmit={handleSubmit(() => {})}>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const, { required: true })}
-                defaultValue={field.value}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-            <button type="button" onClick={() => move(0, 1)}>
-              move
-            </button>
-            <button>submit</button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      await actComponent(async () => {
-        fireEvent.click(screen.getByRole('button', { name: /submit/i }));
-      });
-
-      expect(errors.test[0]).toBeUndefined();
-      expect(errors.test[1]).toBeDefined();
-
-      fireEvent.click(screen.getByRole('button', { name: /move/i }));
-
-      expect(errors.test[0]).toBeDefined();
-      expect(errors.test[1]).toBeUndefined();
-    });
-
-    it('should move touched fields', async () => {
-      let touched: any;
-      const Component = () => {
-        const { register, formState, control } = useForm({
-          defaultValues: { test: [{ value: 'test' }] },
-        });
-        const { fields, append, move } = useFieldArray({
-          control,
-          name: 'test',
-        });
-
-        touched = formState.touched;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-            <button type="button" onClick={() => append({ value: '' })}>
-              append
-            </button>
-            <button type="button" onClick={() => move(0, 1)}>
-              move
-            </button>
-            <button>submit</button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      fireEvent.blur(screen.getAllByRole('textbox')[0]);
-
-      fireEvent.click(screen.getByRole('button', { name: /move/i }));
-
-      expect(touched).toEqual({
-        test: [undefined, { value: true }],
-      });
-    });
-
-    it('should trigger reRender when user is watching the all field array', () => {
-      const watched: any[] = [];
-      const Component = () => {
-        const { register, watch, control } = useForm({
-          defaultValues: {
-            test: [{ value: '1' }, { value: '2' }],
-          },
-        });
-        const { fields, move } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        watched.push(watch());
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                defaultValue={field.value}
-                {...register(`test.${i}.value` as const)}
-              />
-            ))}
-            <button type="button" onClick={() => move(0, 1)}>
-              move
-            </button>
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: 'move' }));
-
-      expect(watched).toEqual([
-        { test: [{ value: '1' }, { value: '2' }] }, // first render
-        { test: [{ value: '1' }, { value: '2' }] }, // render inside useEffect in useFieldArray
-        { test: [{ value: '2' }, { value: '1' }] }, // render inside move method
-        { test: [{ value: '2' }, { value: '1' }] }, // render inside useEffect in useFieldArray
-      ]);
-    });
-
-    it('should populate all fields with default values', () => {
-      let getValues: any;
-      const Component = () => {
-        const { register, control, getValues: tempGetValues } = useForm({
-          defaultValues: {
-            test: [{ value: '1' }, { value: '2' }],
-          },
-        });
-        const { fields } = useFieldArray({
-          control,
-          name: 'test',
-        });
-        getValues = tempGetValues;
-
-        return (
-          <form>
-            {fields.map((field, i) => (
-              <input
-                key={field.id}
-                {...register(`test.${i}.value` as const)}
-                defaultValue={field.value}
-              />
-            ))}
-          </form>
-        );
-      };
-
-      render(<Component />);
-
-      expect(getValues()).toEqual({ test: [{ value: '1' }, { value: '2' }] });
-    });
-
-    it('should return watched value with watch API', async () => {
-      const renderedItems: any = [];
-      const Component = () => {
-        const { watch, register, control } = useForm<{
-          test: {
-            value: string;
-          }[];
-        }>();
-        const { fields, append, move } = useFieldArray({
-          name: 'test',
-          control,
-        });
-        const watched = watch('test');
-        const isMoved = React.useRef(false);
-        if (isMoved.current) {
-          renderedItems.push(watched);
-        }
-
-        return (
-          <div>
-            {fields.map((field, i) => (
-              <div key={`${field.id}`}>
-                <input
-                  defaultValue={field.value}
-                  {...register(`test.${i}.value` as const)}
-                />
-              </div>
-            ))}
-            <button onClick={() => append({ value: '' })}>append</button>
-            <button
-              onClick={() => {
-                move(0, 1);
-                isMoved.current = true;
-              }}
-            >
-              move
-            </button>
-          </div>
-        );
-      };
-
-      render(<Component />);
-
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-      fireEvent.click(screen.getByRole('button', { name: /append/i }));
-
-      const inputs = screen.getAllByRole('textbox');
-
-      fireEvent.change(inputs[0], {
-        target: { name: 'test[0].value', value: '111' },
-      });
-      fireEvent.change(inputs[1], {
-        target: { name: 'test[1].value', value: '222' },
-      });
-
-      fireEvent.click(screen.getByRole('button', { name: /move/i }));
-
-      await waitFor(() =>
-        expect(renderedItems).toEqual([
-          [{ value: '222' }, { value: '111' }],
-          [{ value: '222' }, { value: '111' }],
-        ]),
-      );
-    });
-
-    describe('with resolver', () => {
-      it('should invoke resolver when formState.isValid true', async () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-            defaultValues: {
-              test: [{ value: '1' }, { value: '2' }],
-            },
-          });
-          const { move } = useFieldArray({ control, name: 'test' });
-          return { formState, move };
-        });
-
-        result.current.formState.isValid;
-
-        await act(async () => {
-          result.current.move(0, 1);
-        });
-
-        expect(resolver).toBeCalledWith(
-          {
-            test: [{ value: '2' }, { value: '1' }],
-          },
-          undefined,
-          false,
-        );
-      });
-
-      it('should not invoke resolver when formState.isValid false', () => {
-        const resolver = jest.fn().mockReturnValue({});
-
-        const { result } = renderHook(() => {
-          const { formState, control } = useForm({
-            mode: VALIDATION_MODE.onChange,
-            resolver,
-            defaultValues: {
-              test: [{ value: '1' }, { value: '2' }],
-            },
-          });
-          const { move } = useFieldArray({ control, name: 'test' });
-          return { formState, move };
-        });
-
-        act(() => {
-          result.current.move(0, 1);
-        });
-
-        expect(resolver).not.toBeCalled();
-      });
     });
   });
