+++ src/__tests__/useFieldArray/prepend.test.tsx
@@ -1,0 +1,453 @@
+import { useForm } from '../../useForm';
+import { useFieldArray } from '../../useFieldArray';
+import {
+  act as actComponent,
+  fireEvent,
+  render,
+  screen,
+  waitFor,
+} from '@testing-library/react';
+import { act, renderHook } from '@testing-library/react-hooks';
+import * as React from 'react';
+import { VALIDATION_MODE } from '../../constants';
+import { mockGenerateId } from '../useFieldArray.test';
+
+let nodeEnv: string | undefined;
+
+describe('prepend', () => {
+  beforeEach(() => {
+    mockGenerateId();
+    nodeEnv = process.env.NODE_ENV;
+    process.env.NODE_ENV = 'production';
+  });
+
+  afterEach(() => {
+    jest.resetAllMocks();
+    jest.restoreAllMocks();
+    process.env.NODE_ENV = nodeEnv;
+  });
+
+  it('should pre-append data into the fields', () => {
+    let currentFields: any = [];
+
+    const Component = () => {
+      const { control, register } = useForm<{
+        test: {
+          test: string;
+        }[];
+      }>();
+      const { fields, prepend } = useFieldArray({
+        control,
+        name: 'test',
+      });
+
+      currentFields = fields;
+
+      return (
+        <form>
+          {fields.map((field, index) => (
+            <div key={field.id}>
+              <input
+                {...register(`test.${index}.test` as const)}
+                defaultValue={field.test}
+              />
+            </div>
+          ))}
+          <button type={'button'} onClick={() => prepend({ test: 'test' })}>
+            prepend
+          </button>
+          <button
+            type={'button'}
+            onClick={() =>
+              prepend([{ test: 'test-batch' }, { test: 'test-batch1' }])
+            }
+          >
+            prependBatch
+          </button>
+        </form>
+      );
+    };
+
+    render(<Component />);
+
+    act(() => {
+      fireEvent.click(screen.getByRole('button', { name: 'prepend' }));
+    });
+
+    act(() => {
+      expect(currentFields).toEqual([{ id: '0', test: 'test' }]);
+    });
+
+    act(() => {
+      fireEvent.click(screen.getByRole('button', { name: 'prepend' }));
+    });
+
+    act(() => {
+      expect(currentFields).toEqual([
+        { id: '1', test: 'test' },
+        { id: '0', test: 'test' },
+      ]);
+    });
+
+    act(() => {
+      fireEvent.click(screen.getByRole('button', { name: 'prependBatch' }));
+    });
+
+    act(() => {
+      expect(currentFields).toEqual([
+        { id: '2', test: 'test-batch' },
+        { id: '3', test: 'test-batch1' },
+        { id: '1', test: 'test' },
+        { id: '0', test: 'test' },
+      ]);
+    });
+  });
+
+  it.each(['isDirty', 'dirtyFields'])(
+    'should be dirty when value is prepended with %s',
+    () => {
+      const { result } = renderHook(() => {
+        const { register, formState, control } = useForm();
+        const { fields, prepend } = useFieldArray({
+          control,
+          name: 'test',
+        });
+
+        return { register, formState, fields, prepend };
+      });
+
+      result.current.formState.isDirty;
+      result.current.formState.dirty;
+
+      act(() => {
+        result.current.prepend({ value: 'test' });
+      });
+
+      act(() => {
+        result.current.prepend({ value: 'test1' });
+      });
+
+      act(() => {
+        result.current.prepend({ value: 'test2' });
+      });
+
+      expect(result.current.formState.isDirty).toBeTruthy();
+      expect(result.current.formState.dirty).toEqual({
+        test: [{ value: true }, { value: true }, { value: true }],
+      });
+    },
+  );
+
+  it('should set prepended values to formState.touched', () => {
+    let touched: any;
+
+    const Component = () => {
+      const { register, formState, control } = useForm();
+      const { fields, prepend } = useFieldArray({
+        control,
+        name: 'test',
+      });
+
+      touched = formState.touched;
+
+      return (
+        <form>
+          {fields.map((field, i) => (
+            <input key={field.id} {...register(`test.${i}.value`)} />
+          ))}
+          <button type="button" onClick={() => prepend({ value: `test${1}` })}>
+            prepend
+          </button>
+        </form>
+      );
+    };
+
+    render(<Component />);
+
+    fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
+
+    fireEvent.blur(screen.getAllByRole('textbox')[0]);
+
+    fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
+    fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
+
+    expect(touched).toEqual({
+      test: [undefined, { value: true }, { value: true }],
+    });
+  });
+
+  it('should prepend error', async () => {
+    let errors: any;
+    const Component = () => {
+      const {
+        register,
+        formState: { errors: tempErrors },
+        handleSubmit,
+        control,
+      } = useForm<{
+        test: { value: string }[];
+      }>();
+      const { fields, prepend } = useFieldArray({
+        control,
+        name: 'test',
+      });
+      errors = tempErrors;
+
+      return (
+        <form onSubmit={handleSubmit(() => {})}>
+          {fields.map((field, i) => (
+            <input
+              key={field.id}
+              {...register(`test.${i}.value` as const, { required: true })}
+              defaultValue={field.value}
+            />
+          ))}
+          <button type="button" onClick={() => prepend({ value: '' })}>
+            prepend
+          </button>
+          <button>submit</button>
+        </form>
+      );
+    };
+
+    render(<Component />);
+
+    fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
+
+    expect(errors.test).toBeUndefined();
+
+    await actComponent(async () => {
+      fireEvent.click(screen.getByRole('button', { name: /submit/i }));
+    });
+
+    fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
+
+    expect(errors.test).toHaveLength(2);
+  });
+
+  it('should trigger reRender when user is watching the all field array', () => {
+    const watched: any[] = [];
+    const Component = () => {
+      const { register, watch, control } = useForm<{
+        test: { value: string }[];
+      }>();
+      const { fields, prepend } = useFieldArray({
+        control,
+        name: 'test',
+      });
+      watched.push(watch());
+
+      return (
+        <form>
+          {fields.map((field, i) => (
+            <input
+              key={field.id}
+              {...register(`test.${i}.value` as const)}
+              defaultValue={field.value}
+            />
+          ))}
+          <button type="button" onClick={() => prepend({ value: '' })}>
+            prepend
+          </button>
+        </form>
+      );
+    };
+
+    render(<Component />);
+
+    fireEvent.click(screen.getByRole('button', { name: 'prepend' }));
+
+    expect(watched).toEqual([
+      {}, // first render
+      {}, // render inside useEffect in useFieldArray
+      { test: [{ value: '' }] }, // render inside prepend method
+      { test: [{ value: '' }] }, // render inside useEffect in useFieldArray
+    ]);
+  });
+
+  it('should return watched value with watch API', async () => {
+    const renderedItems: any = [];
+    const Component = () => {
+      const { watch, register, control } = useForm<{
+        test: {
+          value: string;
+        }[];
+      }>();
+      const { fields, append, prepend } = useFieldArray({
+        name: 'test',
+        control,
+      });
+      const watched = watch('test');
+      const isPrepended = React.useRef(false);
+      if (isPrepended.current) {
+        renderedItems.push(watched);
+      }
+      return (
+        <div>
+          {fields.map((field, i) => (
+            <div key={`${field.id}`}>
+              <input
+                defaultValue={field.value}
+                {...register(`test.${i}.value` as const)}
+              />
+            </div>
+          ))}
+          <button onClick={() => append({ value: '' })}>append</button>
+          <button
+            onClick={() => {
+              prepend({ value: 'test' });
+              isPrepended.current = true;
+            }}
+          >
+            prepend
+          </button>
+        </div>
+      );
+    };
+
+    render(<Component />);
+
+    fireEvent.click(screen.getByRole('button', { name: /append/i }));
+    fireEvent.click(screen.getByRole('button', { name: /append/i }));
+
+    const inputs = screen.getAllByRole('textbox');
+
+    fireEvent.input(inputs[0], {
+      target: { name: 'test[0].value', value: '111' },
+    });
+    fireEvent.input(inputs[1], {
+      target: { name: 'test[1].value', value: '222' },
+    });
+
+    fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
+
+    await waitFor(() =>
+      expect(renderedItems).toEqual([
+        [{ value: 'test' }, { value: '111' }, { value: '222' }],
+        [{ value: 'test' }, { value: '111' }, { value: '222' }],
+      ]),
+    );
+  });
+
+  it('should focus if shouldFocus is true', () => {
+    const Component = () => {
+      const { register, control } = useForm<{
+        test: { value: string }[];
+      }>({
+        defaultValues: {
+          test: [{ value: '1' }, { value: '2' }],
+        },
+      });
+      const { fields, prepend } = useFieldArray({ name: 'test', control });
+
+      return (
+        <form>
+          {fields.map((field, i) => (
+            <input
+              key={field.id}
+              {...register(`test.${i}.value` as const)}
+              defaultValue={field.value}
+            />
+          ))}
+          <button type="button" onClick={() => prepend({ value: '' })}>
+            prepend
+          </button>
+        </form>
+      );
+    };
+
+    render(<Component />);
+
+    fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
+
+    const inputs = screen.getAllByRole('textbox');
+
+    expect(inputs).toHaveLength(3);
+    expect(document.activeElement).toEqual(inputs[0]);
+  });
+
+  it('should not focus if shouldFocus is false', () => {
+    const Component = () => {
+      const { register, control } = useForm({
+        defaultValues: {
+          test: [{ value: '1' }, { value: '2' }],
+        },
+      });
+      const { fields, prepend } = useFieldArray({ name: 'test', control });
+
+      return (
+        <form>
+          {fields.map((field, i) => (
+            <input
+              key={field.id}
+              {...register(`test.${i}.value` as const)}
+              defaultValue={field.value}
+            />
+          ))}
+          <button
+            type="button"
+            onClick={() => prepend({ value: '' }, { shouldFocus: false })}
+          >
+            prepend
+          </button>
+        </form>
+      );
+    };
+
+    render(<Component />);
+
+    fireEvent.click(screen.getByRole('button', { name: /prepend/i }));
+    const inputs = screen.getAllByRole('textbox');
+
+    expect(inputs).toHaveLength(3);
+    expect(document.activeElement).toEqual(document.body);
+  });
+
+  describe('with resolver', () => {
+    it('should invoke resolver when formState.isValid true', async () => {
+      const resolver = jest.fn().mockReturnValue({});
+
+      const { result } = renderHook(() => {
+        const { formState, control } = useForm({
+          mode: VALIDATION_MODE.onChange,
+          resolver,
+        });
+        const { prepend } = useFieldArray({ control, name: 'test' });
+        return { formState, prepend };
+      });
+
+      result.current.formState.isValid;
+
+      await act(async () => {
+        result.current.prepend({ value: '1' });
+      });
+
+      expect(resolver).toBeCalledWith(
+        {
+          test: [{ value: '1' }],
+        },
+        undefined,
+        false,
+      );
+    });
+
+    it('should not invoke resolver when formState.isValid false', () => {
+      const resolver = jest.fn().mockReturnValue({});
+
+      const { result } = renderHook(() => {
+        const { formState, control } = useForm({
+          mode: VALIDATION_MODE.onChange,
+          resolver,
+        });
+        const { prepend } = useFieldArray({ control, name: 'test' });
+        return { formState, prepend };
+      });
+
+      act(() => {
+        result.current.prepend({ value: '1' });
+      });
+
+      expect(resolver).not.toBeCalled();
+    });
+  });
+});
